requirements.txt
autopep8==1.4.4
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Cors==3.0.8
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
MarkupSafe==1.1.1
msgpack==0.6.2
mypy==0.761
mypy-extensions==0.4.3
pycodestyle==2.5.0
PyMySQL==0.9.3
PyYAML==5.2
redis==3.3.11
requests==2.22.0
six==1.13.0
typed-ast==1.4.0
typing-extensions==3.7.4.1
urllib3==1.25.7
Werkzeug==0.16.0
